Update to latest internal revision. (#108)
* Remove compiler warnings, mostly by comparing sizes better wrt signedness.

Stop relying on implicit conversions.
Use multiline comments where a backslash is necessary to show perf commandlines.

PiperOrigin-RevId: 399473470

* Add an icon indicating status with a link to Github Actions from our README.

PiperOrigin-RevId: 400840383

* Add {code,data}_page_size label support

PiperOrigin-RevId: 401781351

* Fix typo

PiperOrigin-RevId: 404883340

+++ src/perf_data_handler_test.cc
@@ -96,7 +96,7 @@ class TestPerfDataHandler : public PerfDataHandler {
       _seen_addr_mappings.push_back(nullptr);
     }
     EXPECT_EQ(_expected_branch_stack.size(), sample.branch_stack.size());
-    for (int i = 0; i < sample.branch_stack.size(); i++) {
+    for (size_t i = 0; i < sample.branch_stack.size(); i++) {
       CheckBranchEquality(_expected_branch_stack[i], sample.branch_stack[i]);
     }
   }
